[#16] Use assoc array when defining scripts in PHP (#17)
If a sequential array is used for declaring the scripts to be deferred
in PHP, that PHP array is converted to a JavaScript array, and then
cannot be properly extended without possibly duplicating and / or
removing existing deferred scripts, in the case of an AJAX request that
extends the Drupal settings.

Using an associative array in PHP will convert to a JavaScript object.
In this commit, we declare the associative array with equivalent keys
and values, which will ensure that all scripts are included, and that no
duplicates are possible. We then use `Object.keys()` to convert that
object into a JavaScript array just before processing with LazyLoad.